Fix Server Crash (with a single modify file)
I had launched a PR earlier this week fixing the server crash problem.
I had actually uploaded all the files from my 1.16.5-2.39 sdk which created many conflicts with the branch. As well as the wrong targeting of the branch.

Here is a version that modifies only one file but still fixes the server crash.